BackwardsCompatibilityBreak - fORM::addCustomClassTableMapping() was renamed to fORM::mapClassToTable()
Fixed tickets #148 and #265 - Added the ability to set a database name via fORMDatabase::attach() and fORMSchema::attach() which can then be used to support multiple databases for the ORM via fORM::mapClassToDatabase(). In addition, different databases can be set for read and write operations in fORMDatabase::attach() and fORMSchema::attach(), which allows for master/slave setups.

git-svn-id: http://svn.flourishlib.com/trunk@726 8e09ec71-114c-0410-a285-86ecef5c4c40