BackwardsCompatibilityBreak - fRequest::get() now strips out low bytes characters if no `$cast_to`, or if a `string` or `array` `$cast_to` is specified. Using the new `binary` `$cast_to` will leave all bytes intact.
Completed ticket #486 - added the `binary` type to fRequest::get().

Completed ticket #533 - fRequest::get() will cast all strings of digits in the range of PHP integer type to such an integer when the `integer` or `int` `$cast_to` is specified. If the integer is beyond the range, a string of the digits will be returned. fRequest::get() can be forced to always return a true integer by passing `integer!`.

git-svn-id: http://svn.flourishlib.com/trunk@943 8e09ec71-114c-0410-a285-86ecef5c4c40